requirements.txt
Babel>=0.9.6
eventlet>=0.13.0
oslo.messaging>=1.3.0a4
pbr>=0.5.21,<1.0
requests>=1.1

kombu>=2.4.8
